+++ functions.php
@@ -268,8 +268,8 @@ function parse_ACF_block( $blockname ) {
 //11 = hidden
 //13 = page_home
 function pre_get_posts_filtered( $query ) {
-	if ( ! is_admin() && $query->is_main_query() ) {
-			$query->set( 'cat', '-11' );
+	if ( $query->is_main_query() ) {
+			$query->set( 'category__not_in', 'hidden' );
 	}
 }
 add_action( 'pre_get_posts', 'pre_get_posts_filtered' );