+++ page-search.php
@@ -15,30 +15,9 @@ get_header();
 		<i></i>
 	</div>
 	<div class="projects">
-		<?php
-		// get all the posts
-			$all_posts = get_posts(array(
-				'post_type' => 'post',
-				'posts_per_page'   => -1,
-				)
-		);
-
-		if (count($all_posts) > 1):
-			foreach($all_posts as $post):
-		?>
-			<h2 class="project-line">
-				<a href="<?php the_permalink(); ?>"
-					title="<?php the_title(); ?>"
-					alt="Project of Tabea Nixdorff: <?php the_title(); ?>"
-					data-tag="<?php if ($posttags):
-						foreach($posttags as $tag): echo $tag->name . ' ';
-						endforeach; endif;?>"
-				>
-				<?php the_title(); ?>
-				<?php parse_ACF_block( 'acf/project' ); ?>
-				</a>
-			</h2>
-	<?php endforeach; endif; ?>
+		<?php get_template_part( 'template-parts/loops/projects', NULL, array(
+			'set-data-tag' => false
+		) ); ?>
 	</div><!-- end of .projects -->